refactor: Update project structure This commit updates the project structure to improve organization and maintainability.
**Changes:**
- Moved `data-access` folder to a `home` directory
- Relocated `contact` and `search-contact` components  to a `home/ui`` directory
- Updated import paths in affected files
