+++ src/app/app.routes.ts
@@ -1,6 +1,6 @@
 import { Routes } from '@angular/router';
 import { HomeComponent } from './home/home.component';
-import { ContactComponent } from './shared/ui/components/contact/contact.component';
+import { ContactComponent } from './home/ui/components/contact/contact.component';
 import { Error404Component } from './error404/error404.component';
 
 export const routes: Routes = [
